views (wip)

+++ src/Aardvark.Geometry.PointSet/Octrees/PointSet.cs
@@ -57,7 +57,7 @@ namespace Aardvark.Geometry.Points
         /// <summary>
         /// Creates pointset from given root cell.
         /// </summary>
-        public PointSet(Storage storage, string key, Guid? rootCellId, long splitLimit)
+        public PointSet(Storage storage, string key, Guid? rootCellId, int splitLimit)
         {
             Storage = storage;
             Id = key ?? throw new ArgumentNullException(nameof(key));
@@ -118,7 +118,7 @@ namespace Aardvark.Geometry.Points
 
         /// <summary>
         /// </summary>
-        public long SplitLimit { get; }
+        public int SplitLimit { get; }
         
         /// <summary>
         /// </summary>
